Cleanup data controller caching code
