+++ packages/ag-charts-community/src/chart/chart.ts
@@ -38,7 +38,8 @@ import { ChartContext } from './chartContext';
 import { ChartHighlight } from './chartHighlight';
 import type { ChartMode } from './chartMode';
 import { ChartUpdateType } from './chartUpdateType';
-import { type CachedData, DataController } from './data/dataController';
+import { type CachedData } from './data/caching';
+import { DataController } from './data/dataController';
 import { axisRegistry } from './factory/axisRegistry';
 import { EXPECTED_ENTERPRISE_MODULES } from './factory/expectedEnterpriseModules';
 import { legendRegistry } from './factory/legendRegistry';
